requirements.txt
Flask==2.0.2
tensorflow==2.7.0
requests
NumPy
pillow